lean in to the theme w/ theme-color & selection
Modern browsers support `theme-color` which can be used to change the
title bar of the of the browser. This is common on Android’s default
browsers and Vivaldi, but some web extensions like VivaldiFox.
`::selection` the text selection color when dragging over content. In
both of these cases, I chose the horse head’s mane’s red to contrast the
more-common orange color on the site (used on hyperlinks) which helps stikes
the 50/50 balance of the logo’s two dominate colors. (without `theme-color`
VivaldiFox picked the brown of the pole which just doesn’t look nice).

I believe these changes help unify the aesthetic.

I also removed an unused CSS variable changed the red in the dark theme to
match the main (I think there may have been an oversight whereas the orange
on light theme is darker to provide better contrast).